Remove redundant spatial test casts

+++ packages/libmylite/tests/runtime_spatial_bounded_geometry_nesting_test.c
@@ -608,7 +608,7 @@ static int expect_direct_success(
 
     if (rc == 0) {
         failures +=
-            mylite_test_expect_int((int)result.kind, (int)MYLITE_SPATIAL_RESULT_GEOMETRY, context);
+            mylite_test_expect_int((int)result.kind, MYLITE_SPATIAL_RESULT_GEOMETRY, context);
     }
     mylite_spatial_result_deinit(&result);
     return failures;
@@ -631,7 +631,7 @@ static int expect_direct_error(
 
     failures += mylite_test_expect_int(error.code, expected_code, context);
     failures += mylite_test_expect_text(error.sqlstate, expected_sqlstate, context);
-    failures += mylite_test_expect_int((int)result.kind, (int)MYLITE_SPATIAL_RESULT_NULL, context);
+    failures += mylite_test_expect_int((int)result.kind, MYLITE_SPATIAL_RESULT_NULL, context);
     mylite_spatial_result_deinit(&result);
     return failures;
 }
